Loading page that retrieves user data from the api
